@@ -1,5 +1,5 @@
 import React from "react";
-import { Switch, Route } from "react-router-dom";
+import { Routes, Route } from "react-router-dom";
 import Calculator from "components/calculator";
 import Results from "components/results";
 import Header from "components/header";
@@ -11,10 +11,10 @@ function App() {
       <Header />
       <main className="max-w-screen-sm mx-auto px-4 md:px-0" style={{paddingLeft: "calc(100vw - 100%)" }}>
         <h1 className="text-2xl mt-6 mb-8">FFXIV 陆行鸟染色计算器</h1>
-        <Switch>
-          <Route exact path="/" render={(props) => <Calculator />} />
-          <Route exact path="/:from/:to" render={(props) => <Results/>} />
-        </Switch>
+        <Routes>
+          <Route path="/" element={<Calculator />} />
+          <Route path="/:from/:to" element={<Results/>} />
+        </Routes>
       </main>
       <footer className="text-center text-sm text-gray-500 my-2">
         <p>最终幻想14陆行鸟染色计算器<span className="px-1">·</span>
